Iss #12: Implementing `MySQLConnection::insert()` (tested)
